+++ test/tutorials/src/server.ts
@@ -21,6 +21,9 @@ import AsyncTestUtil from 'async-test-util';
 import * as request from 'request-promise-native';
 import * as assert from 'assert';
 
+import * as os from 'os';
+import * as path from 'path';
+
 async function run() {
 
 
@@ -59,7 +62,12 @@ async function run() {
     const serverState = db.server({
         path: '/db',
         port: 3000,
-        cors: true
+        cors: true,
+        pouchdbExpressOptions: {
+            inMemoryConfig: true, // do not write a config.json
+            // save logs in tmp folder
+            logPath: path.join(os.tmpdir(), 'rxdb-server-log.txt')
+        }
     });
     console.log('You can now open http://localhost:3000/db');
     // and should see something like '{"express-pouchdb":"Welcome!","version":"4.1.0","pouchdb-adapters":["memory"],"vendor":{"name":"PouchDB authors","version":"4.1.0"},"uuid":"b2de36bf-7d4f-4ad1-89a4-da08ec0de227"}'
@@ -71,7 +79,7 @@ async function run() {
     // check access to path
     const gotJson = await request(colUrl);
     const got = JSON.parse(gotJson);
-    assert.equal(got.doc_count, 1);
+    assert.strictEqual(got.doc_count, 1);
 
     /**
      * on the client